requirements.txt
fastapi
uvicorn
requests
PyPDF2
python-docx
sentence-transformers
faiss-cpu
python-dotenv
torch --index-url https://download.pytorch.org/whl/cpu
